pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "local-llm-training-k8s"
version = "0.1.0"
description = "Training pipeline for distributed decoder-only transformer on local Kubernetes"
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = { file = "LICENSE" }
authors = [{ name = "Gabriel Levaillant" }]
keywords = ["llm", "training", "kubernetes"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
    "pydantic>=2.0.0,<3.0.0",
    "pyyaml>=6.0,<7.0.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.9",
  "mypy>=1.13.0",
  "pytest>=8.3.3",
  "pre-commit>=3.8.0",
  "types-pyyaml>=6.0",
]

[project.scripts]
llmtrain = "llmtrain.cli:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.11"
mypy_path = ["src"]
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
explicit_package_bases = true
namespace_packages = true

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
branch = true
source = ["llmtrain"]

[tool.coverage.report]
show_missing = true
skip_covered = true

